added

+++ dot_hardware/motor.h
@@ -99,11 +99,11 @@ int Motor::control(){
     if(val<0){
         digitalWrite(motor_p, HIGH);
         digitalWrite(motor_n, LOW);
-        softPwmWrite(motor_e, min(-val, m_effort));
+        softPwmWrite(motor_e, (int)min(-val, m_effort));
     } else {
         digitalWrite(motor_p, LOW);
         digitalWrite(motor_n, HIGH);
-        softPwmWrite(motor_e, min(val, m_effort));
+        softPwmWrite(motor_e, (int)min(val, m_effort));
     }
     return 1;
 }
